requirements.txt
flask
flask-cors
faster-whisper
torch>=2.0
torchaudio
pyannote.audio==3.1
transformers  # Needed by pyannote.audio
numpy         # Sometimes needed for audio models
scipy         # Safer for audio I/O processing
huggingface-hub  # Required for model downloading
ctranslate2